+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.17.0
+
+- Add `Msix.assetUri()` to get an `ms-appx:///` URI out of any Flutter asset name.
+- Add `Msix.hasPackageIdentity()` to check if the currently running program was installed with an MSIX.
+
 ## 3.16.8
 
 - update VCLibs files [#273](https://github.com/YehudaKremer/msix/pull/273)